db: rework the mechanism to support a changing rangeDelIter
When using a `levelIter`, the `rangeDelIter` changes with each file.
The current mechanism for higher levels (in particular the merging
iterator) to obtain the current `rangeDelIter` is to configure the
`levelIter` with a pointer to a field. This is very subtle, as the
field seems to "magically" change from the perspective of the higher
level code.

This commit changes this mechanism: we define a `rangeDelIterHolder`
which can be configured either with a fixed `rangeDeliter` or with a
`levelIter` that is configured to expose the current `rangeDelIter`. A
`get()` method returns either the fixed `rangeDelIter` or the
`levelIter`'s current `rangeDelIter`.