db: improve delete pacing
We improve the delete pacer to be able to scale delete pacing instead
of making an on/off decision.

The new pacer uses low/high thresholds for various factors; on one
side of the threshold we pace at the target rate and on the other side
we don't throttle at all. In-between we scale the wait time per byte
linearly.

Thresholds are as follows:

|                              | Start increasing rate at | Disable pacing at |
| ---------------------------- | ------------------------ | ----------------- |
| Free space                   | 32GB                     | 16GB              |
| Obsolete to live bytes ratio | 5%                       | 20%               |
| Obsolete bytes (*new*)       | 1GB                      | 10GB              |

Fixes #2662.