github: rework nightly workflows and run on crl-release-25.2
This change reworks the nightly workflows so they can be run on any
branch. We now have a "master" workflow that calls all the nightly
workflows (instead of having each one get scheduled separately).

They run on master and crl-release-25.2. Note that older branches run
most of the tests in PRs.